Константа пустой точки маршрута вынесена в отдельный файл
